Added the option to restrict clozing to whole words only.  Useful for languages that don't use kanji.
For example, given the following text:
"I take nothing for granted, especially not breakfast."

suppose we want to cloze the word "not".  With the "Whole words only" box unchecked, the word "not" and the first three letters of
"nothing" will be clozed.  If the box is checked however, only the whole word "not" will be clozed.